requirements.txt
certifi==2024.8.30
cffi==1.17.1
chardet==3.0.4
cryptography==43.0.3
decorator==5.1.1
et_xmlfile==2.0.0
future==1.0.0
idna==2.8
numpy==2.1.3
openpyxl==3.1.5
pandas==2.2.3
pycparser==2.22
pyfiglet==0.8.post1
python-dateutil==2.9.0.post0
python-dotenv==0.17.1
python-telegram-bot==12.7
pytz==2024.2
requests==2.22.0
schedule==1.2.2
six==1.16.0
tornado==6.4.1
tzdata==2024.2
urllib3==1.25.11
sqlalchemy==2.0.36